+++ pytket/extensions/qiskit/backends/ibmq_emulator.py
@@ -129,9 +129,7 @@ class IBMQEmulatorBackend(Backend):
     ) -> List[ResultHandle]:
         """
         See :py:meth:`pytket.backends.Backend.process_circuits`.
-        Supported kwargs: `seed`, `postprocess`, `seed_auto_increase`.
-        seed_auto_increase=True will automatically increase the seed by one for the
-        different batches when more than one circuit is submitted
+        Supported kwargs: `seed`, `postprocess`.
         """
         circuits = list(circuits)
         n_shots_list = Backend._get_n_shots_as_list(
@@ -177,7 +175,7 @@ class IBMQEmulatorBackend(Backend):
                 options.execution.shots = n_shots
                 options.simulator.noise_model = self._noise_model
                 options.seed_simulator = seed
-                if kwargs.get("seed_auto_increase") and type(seed) is int:
+                if type(seed) is int:
                     seed += 1
                 sampler = Sampler(session=self._session, options=options)
                 job = sampler.run(circuits=qcs)